Markdown関連の変換処理をutilにまとめた
